refactor(threads): useThreadMessaging 外移 memoryPick 模块级 helper 与类型定义（4025→3756）
